+++ src/common/components/footer/CopyRight.jsx
@@ -1,0 +1,5 @@
+const CopyRight = ({ copyright }) => {
+  return <p className="text-gray-600">{copyright}</p>
+}
+
+export default CopyRight
