+++ src/common/components/footer/Footer.jsx
@@ -1,10 +1,10 @@
-import React from 'react'
+import CopyRight from '@common/components/footer/CopyRight.jsx'
 
 const Footer = () => {
   return (
     <footer className="text-center">
       <div className="container mx-auto text-center pt-16 px-10 pb-10">
-        <p className="text-gray-600">&copy; 2024 HYEHYE blog.</p>
+        <CopyRight copyright={'Copyright © 2024 HYEHYE blog.'} />
       </div>
     </footer>
   )
