+++ src/router.jsx
@@ -3,8 +3,8 @@ import { createBrowserRouter, Outlet } from 'react-router-dom'
 
 // common
 import Layout from '@layouts/Layout.jsx'
-import Loading from '@common/components/Loading.jsx'
-import NotFound from '@common/components/NotFound.jsx'
+import Loading from '@common/components/etc/Loading.jsx'
+import NotFound from '@common/components/etc/NotFound.jsx'
 import NewWrite from '@common/components/NewWrite.jsx'
 
 // pages
